+++ src/Doc/methods/output/01-text.js
@@ -1,3 +1,9 @@
+const shouldTrim = {
+  clean: true,
+  reduced: true,
+  root: true,
+}
+
 /** return the document as text */
 exports.text = function(options) {
   options = options || {}
@@ -17,9 +23,15 @@ exports.text = function(options) {
     })
   }
 
-  return this.list.reduce((str, p, i) => {
+  let txt = this.list.reduce((str, p, i) => {
     const trimPre = !showFull && i === 0
     const trimPost = !showFull && i === this.list.length - 1
     return str + p.text(options, trimPre, trimPost)
   }, '')
+
+  // clumsy final trim of leading/trailing whitespace
+  if (shouldTrim[options] === true || options.reduced === true || options.clean === true || options.root === true) {
+    txt = txt.trim()
+  }
+  return txt
 }